+++ tool/test/testdata/domain_completeness/valid/packages/dart_sysinfo/lib/src/core/sys_info.dart
@@ -1,0 +1,3 @@
+abstract class SysInfo {
+  CpuDomain get cpu;
+}